Padronizar formato da moeda
Cria e usa no sistema uma classe para padronizar o formato de casas
decimais que usaremos nos atributos.

+++ laravel/database/migrations/2023_11_30_000106_create_transacoes_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->integer('conta_id');
             $table->char('forma_pagamento', 1);
-            $table->float('valor');
+            $table->decimal('valor', 12, 2);
         });
     }
